[DEMO-2557] Copy over workflow & needed files (#6)
* [DEMO-2557] Copy over workflow & needed files

* [DEMO-2557] Add github run id to user config name

* This lets multiple invocations of the workflow run without conflicting
with each other -- the resources will have different names since the run
ids will be different.

* [DEMO-2557] Fix test endpoints, add in retry

* Change the endpoints to actually test the endpoints we want.
* Added in retry gem so that we can wait for the api to be functional
before we start testing. In practice, it looks like the tests are
running while the api is starting up, so giving the tests a way to wait
for the api.

* [DEMO-2557] Update main README with link to UAT documentation

* [DEMO-2557] Add lambda scenario

* Adding in the lambda scenario, deployment, and tests.

* [DEMO-2557] Shorten user credential name

* Deployment was failing to name length limit being exceeded, so removed
part of the name to get the length down.

* [DEMO-2557] Modify workflow to be able to skip matrix combinations

* Certain combinations are currently not possible. Example: deployment
of lambda to azure. The workflow is currently expecting all possible
combinations to be executable, so this commit adds in a way for the
workflow to skip this, by first checking if a particular combination has
a deployment config or not.

* [DEMO-2557] Remove behaviors endpoint test

* Removing behaviors endpoint test. When deployed as a lambda, this
endpoint is not supported.