[DEMO-2561] Code cleanup & bug fixes
* Furthered cleaned up dependencies & initialization in application
entrypoint.
* Renamed database_connection -> database_connector, and mysql as well.
This is to be accurate, since they aren't connections, but can generate
connections.
* Fixed a bug in database inventory repository where on a fresh
deployment, inventory api would fail. This was because the db connection
was made on class initialization and the database is not setup yet.
Moved the connection creation into the repository, to defer connection
creation. Additionally, moved database setup higher within the
application entrypoint.