[DEMO-2561] Add database classes and DI container
* Added classes for interacting with the database, setting up initial
database state, and ancillary concerns.
* Added FileInventoryRepository which will replace the Datastore class
we currently have. The goal with this class is to choose either it or
the database equivalent at application startup.
* Added the DI container class to organize object creation, and provide
a central class for this responsibility overall. Dependency injection is
a useful thing in practice, and a class which makes this process easier
just amplifies that.