Demo 2561 add database (#7)
* [DEMO-2561] Add new requirements

* Added autopep8 for development
* Added mysql-connector-python and dependency-injector for upcoming
changes.

* [DEMO-2561] Add method to app config to return config as dict

* Add method to return app config as dict. This will allow us to import
the app config into the DI container configuration which will be added
in a subsequent commit.
* Also removed unused imports

* [DEMO-2561] Add database classes and DI container

* Added classes for interacting with the database, setting up initial
database state, and ancillary concerns.
* Added FileInventoryRepository which will replace the Datastore class
we currently have. The goal with this class is to choose either it or
the database equivalent at application startup.
* Added the DI container class to organize object creation, and provide
a central class for this responsibility overall. Dependency injection is
a useful thing in practice, and a class which makes this process easier
just amplifies that.

* [DEMO-2561] Update application entrypoint

* Updated the application entrypoint to consume the recent changes. This
primarily involves using the DI container to get dependencies, and
conditionally setting up the database at startup.
* Also removed unused imports.

* [DEMO-2561] Cleanup application entrypoint

* Add more dependencies to the DI container.
* Use DI container to retrieve all dependencies we were previously
creating.
* Improve the database_setup condition. Hopefully its clearer now.
* Remove unused imports.

* [DEMO-2561] Code cleanup & bug fixes

* Furthered cleaned up dependencies & initialization in application
entrypoint.
* Renamed database_connection -> database_connector, and mysql as well.
This is to be accurate, since they aren't connections, but can generate
connections.
* Fixed a bug in database inventory repository where on a fresh
deployment, inventory api would fail. This was because the db connection
was made on class initialization and the database is not setup yet.
Moved the connection creation into the repository, to defer connection
creation. Additionally, moved database setup higher within the
application entrypoint.

* [DEMO-2561] Add database config values to config file on deployment

* [DEMO-2561] Swap check to evaluate '' rather than None

* [DEMO-2561] Add SimpleContainer, use in lambda entrypoint

* Created SimpleContainer, an attempt at copying dependency_injector's
Container class, for use in the lambda entrypoint. This is because the
dependency_injector module is creating runtime errors when used (due
to some special setup needed for installing it).
* Removed unnecessary usings.

* testing out non di framework for lambda entrypoint

* [DEMO-2561] Update UATs

* Add a UAT for each cloud provider x database combination.
* Update workflow to run the database scenario.

* [DEMO-2561] Remove datastore, replace test for FileInventoryRepository

* Since we no longer use / need the datastore class, remove it. Replace
the test written for it to use FileInventoryRepository instead.

* [DEMO-2561] Add unit tests

* Adding in unit tests for MySqlConnector and DatabaseConnectionInfo.

* [DEMO-2561] Add dataclasses to linux-requirements

* Figured out the problem with azure (and probably gcp). The version of
python is < 3.7, which is when dataclasses is available by default. In
order to use dataclasses in python3.6, need to use the backport module,
so I added it to the requirements.

* add missing mysql requirement

* [DEMO-2561] Add database health check endpoint and UAT

* [DEMO-2561] Delete simple container

* [DEMO-2561] Add some comments, context

* [DEMO-2561] Update documentation

* [DEMO-2561] Remove setup_database_method, replace with class

* Created new class SetupDatabaseAction, to be a single place for the
functionality of setting up the database. This lets me remove the 2x
methods I previously had (one for lambda, one for tron). Reduces code
duplication.

* [DEMO-2561] Deduplicate common methods in lambda, tron entrypoint

* Reducing code duplication by moving the get_inventory_data &
selector methods into a single location, which both entrypoints refer
to.
* Also cleaned up imports.

* [DEMO-2561] Update database.database to database.name

* In the application config, there is a field -- database --
nested under database, which is quite confusing.
Swapping the most nested field to name, for clarity and consistency.

* [DEMO-2561] Switch print statements to AppLogging

* Instead of using print, use the AppLogging module.